Uploaded the updated code
There might be confusion. I regularly committed my changes on Jupyter Lab. If that information is not available somehow through GitHub. I had installed the GitHub Jupyter Lab extension and was under the impression that the commits I was making in the Jupyter Lab terminal were being reflected in GitHub as well. If this is not the case I'm very sorry, and am happy to investigate a way to let you access the Version Control documentation for this project. Thank you for your understanding.